lots of stuff
Changed RS, ROB, and free pool to local variables, added wires so hopefully writeback and other things bypassing the pipeline buffer work. Also tried doing retire, doesn't seem to be working but the core of the code should be there. Overall, something seems to be wrong with adding to the RS after instructions get erased from it, but it's compiling at least